packages/alisa-network/pyproject.toml
[project]
name = "alisa-network"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alisa-log",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "websockets>=16.0",
]

[tool.uv.sources]
alisa-log = { workspace = true }
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
